Add models test
