+++ .gitignore
@@ -13,6 +13,7 @@
 composer.phar
 /vendor/
 index.php
+.env
 
 # Commit your application's lock file https://getcomposer.org/doc/01-basic-usage.md#commit-your-composer-lock-file-to-version-control
 # You may choose to ignore a library lock file http://getcomposer.org/doc/02-libraries.md#lock-file
